requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
commonmark==0.9.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
Pygments==2.12.0
pyTelegramBotAPI==4.4.1
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
requests==2.27.1
rich==12.2.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
